updating ux

+++ backend/src/main/java/com/tennis/pricing/PricingService.java
@@ -259,7 +259,7 @@ HOLIDAY:
                                 p.getCourt().getNumber(),
                                 p.getDate(),
                                 p.getStartTime(),
-                                p.getDiscountedPrice()
+                                (double) p.getDiscountedPrice()
                         )
         );
